Kafak topics created programmatically

+++ app-config-data/src/main/java/com/jdutton/poc/config/RetryConfigData.java
@@ -12,7 +12,7 @@ public class RetryConfigData {
 	
 	private Long initialIntervalMs;
 	private Long maxIntervalMs;
-	private Double multiplier;
+	private Integer multiplier;
 	private Integer maxAttempts;
 	private Long sleepTimeMs;
 }
